refactor: [EL-0000] Created new FSD component for grouped radio buttons
