+++ src/[fsd]/features/pipelines/flow-editor/ui/settings/PipelineScheduleModal.jsx
@@ -6,10 +6,9 @@ import 'react-js-cron/dist/styles.css';
 import { Box, Button, GlobalStyles, Typography } from '@mui/material';
 
 import { validateCronExpression } from '@/[fsd]/features/toolkits/indexes/lib/helpers/indexSchedule.helpers.js';
-import { Modal } from '@/[fsd]/shared/ui';
+import { Checkbox, Modal } from '@/[fsd]/shared/ui';
 import InfoTooltip from '@/[fsd]/shared/ui/tooltip/InfoTooltip';
 import FormInput from '@/components/FormInput';
-import RadioButtonGroup from '@/components/RadioButtonGroup';
 
 // Default cron: every Saturday at midnight
 const PipelineCronDefault = '0 0 * * 6';
@@ -60,7 +59,7 @@ const PipelineScheduleModal = props => {
               </Typography>
 
               <Box sx={styles.inputWrapper}>
-                <RadioButtonGroup
+                <Checkbox.RadioButtonGroup
                   label="Schedule Type"
                   value={cronType}
                   items={[
