+++ src/[fsd]/shared/ui/checkbox/index.js
@@ -1,1 +1,2 @@
 export { default as BaseCheckbox } from './BaseCheckbox';
+export { default as RadioButtonGroup } from './RadioButtonGroup';
